ДЗ №6. Применил связку сервис+приемник
ДЗ №6. Применил связку сервис+приемник

+++ app/src/main/java/com/example/k_01/repository/dto/PartDTO.kt
@@ -21,7 +21,7 @@ data class PartDTO(
     @SerializedName("polar")
     val polar: Boolean,
     @SerializedName("prec_mm")
-    val prec_mm: Int,
+    val prec_mm: Double,
     @SerializedName("prec_period")
     val prec_period: Int,
     @SerializedName("prec_prob")
